Text/Banner for files selected/presented on Browse, UI infrastructure, selected files edits.
