+++ src/encoded/static/components/browse/components/index.js
@@ -5,3 +5,4 @@ export { SortController } from './SortController';
 export { SelectedFilesController } from './SelectedFilesController';
 export { CustomColumnController, CustomColumnSelector } from './CustomColumnController';
 export { SearchResultTable, defaultColumnBlockRenderFxn, extendColumnDefinitions, defaultColumnDefinitionMap, columnsToColumnDefinitions } from './SearchResultTable';
+export { AboveTableControls } from './AboveTableControls';
