+++ src/encoded/static/scss/encoded/modules/_tables.scss
@@ -1540,6 +1540,11 @@ $search-results-edge-shadow-color : rgba(0,0,0,0.125);
 		}
 		
 	}
+
+	.box {
+		min-height: 33px;
+		line-height: 33px;
+	}
 }
 
 .result-table {
